requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
Django==1.8.3
django-celery==3.1.16
django-localflavor==1.1
djangorestframework==3.1.3
djangorestframework-jwt==1.7.1
httplib2==0.9.1
kombu==3.0.26
oauthlib==1.0.3
PyJWT==1.4.0
PySocks==1.5.6
pytz==2015.4
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
tweepy==3.4.0
twilio==4.5.0
wheel==0.24.0
